port_samples/obo-authorization/requirements.txt
python-dotenv
microsoft-agents-activity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal